move getFontSizes to text_measurements
